refactor: 공통으로 쓰이는 API, type common 폴더로 이동 #183
