+++ frontend/src/common/types/Specialty.ts
